requirements.txt
django
gunicorn
djangorestframework
django-environ
bleach
python-gnupg